+++ packages/pymodaq/src/pymodaq/utils/custom_ext.py
@@ -3,14 +3,14 @@ from typing import Union, TYPE_CHECKING
 
 from qtpy import QtCore, QtWidgets
 
-from managers.h5manager import H5Manager
+from h5modules.saving import H5Saver
 from pymodaq_utils.enums import StrEnum
 
 from pymodaq_gui.utils import CustomApp, DockArea
 
 from pymodaq.utils.managers.modules.modules_manager import ModulesManager
 from pymodaq.utils.h5modules.module_saving import ModuleSaver
-from pymodaq.utils.managers.h5manager import H5ManagerExt
+
 
 if TYPE_CHECKING:
     from pymodaq.dashboard import DashBoard
@@ -34,7 +34,7 @@ class CustomExt(CustomApp):
                  dashboard: 'DashBoard',
                  module_manager_class=ModulesManager,
                  **kwargs):
-        super().__init__(parent, h5manager_class=H5ManagerExt, **kwargs)
+        super().__init__(parent, **kwargs)
 
         self.dashboard = dashboard
 
@@ -51,9 +51,23 @@ class CustomExt(CustomApp):
         else:
             self._modules_manager = None
 
+        self._module_and_data_saver: ModuleSaver = None  # to use only if you want to save data with
+        #ordered with groups and with control modules, then call self.module_and_data_saver property
+        # a bit complex to use but properly save things from control modules
+
     @property
-    def h5_manager(self) -> H5ManagerExt:
-        return super().h5_manager
+    def module_and_data_saver(self):
+        """ Complex but Standardized way to save data from the CustomExt and
+        especially from the ControlModules"""
+        if (self._module_and_data_saver.h5saver is None
+                or not self._module_and_data_saver.h5saver.isopen()):
+            self._module_and_data_saver.h5saver = self.h5saver
+        return self._module_and_data_saver
+
+    @module_and_data_saver.setter
+    def module_and_data_saver(self, mod: ModuleSaver):
+        self._module_and_data_saver = mod
+        self._module_and_data_saver.h5saver = self.h5saver
 
     def stop(self):
         """ Programmatic method to stop any action in the extension
